+++ core/src/ashlified/entitycomponentsystem/components/SpriterModelComponent.java
@@ -1,12 +1,8 @@
 package ashlified.entitycomponentsystem.components;
 
-import ashlified.graphics.spriterutils.PlayerListenerAdapter;
+import ashlified.graphics.spriterutils.SpriterAnimationController;
 import com.badlogic.ashley.core.Component;
 import com.badlogic.gdx.utils.Pool;
-import com.badlogic.gdx.utils.Queue;
-import com.brashmonkey.spriter.Animation;
-import com.brashmonkey.spriter.Mainline;
-import com.brashmonkey.spriter.Player;
 
 /**
  * Created by karkoon on 25.03.17.
@@ -26,77 +22,5 @@ public final class SpriterModelComponent implements Component, Pool.Poolable {
     @Override
     public void reset() {
         spriterAnimationController = null;
-    }
-
-    public enum AnimationID {
-        IDLE(0), ATTACK(1), DIE(2), DAMAGED(3);
-
-        private int id;
-
-        AnimationID(int id) {
-            this.id = id;
-        }
-
-        public int getId() {
-            return id;
-        }
-    }
-
-    public static class SpriterAnimationController {
-
-        private Player player;
-        private Queue<Runnable> tasks = new Queue<>();
-
-        public Player getPlayer() {
-            return player;
-        }
-
-        private boolean locked = false;
-
-        public void setPlayer(final Player player) {
-            this.player = player;
-            this.player.addListener(new PlayerListenerAdapter() {
-                @Override
-                public void animationFinished(Animation animation) {
-                    if (tasks.size > 0) {
-                        tasks.removeFirst().run();
-                    } else if (animation.id == AnimationID.DIE.getId()) {
-                        locked = false;
-                    } else {
-                        player.setAnimation(AnimationID.IDLE.getId());
-                        locked = false;
-                    }
-                }});
-        }
-
-        public void attackEvent() {
-            animationEvent(AnimationID.ATTACK);
-        }
-
-        public void dieEvent() {
-            animationEvent(AnimationID.DIE);
-        }
-
-        public void damagedEvent() {
-            animationEvent(AnimationID.DAMAGED);
-        }
-
-        private void animationEvent(final AnimationID state) {
-            if (player.getAnimation().id == AnimationID.IDLE.getId()) {
-                player.setAnimation(state.getId());
-                locked = true;
-            } else {
-                tasks.addLast(new Runnable() {
-                    @Override
-                    public void run() {
-                        player.setAnimation(state.getId());
-                    }
-                });
-            }
-        }
-
-        public boolean isLocked() {
-            return locked;
-        }
     }
 }
