+++ core/src/ashlified/entitycomponentsystem/entitysystems/AttackSystem.java
@@ -1,10 +1,13 @@
 package ashlified.entitycomponentsystem.entitysystems;
 
 import ashlified.entitycomponentsystem.components.*;
+import ashlified.graphics.spriterutils.AnimationID;
+import ashlified.graphics.spriterutils.SpriterAnimationController;
 import com.badlogic.ashley.core.ComponentMapper;
 import com.badlogic.ashley.core.Entity;
 import com.badlogic.ashley.core.Family;
 import com.badlogic.ashley.systems.IteratingSystem;
+import com.brashmonkey.spriter.Animation;
 
 public class AttackSystem extends IteratingSystem {
 
@@ -26,17 +29,21 @@ public class AttackSystem extends IteratingSystem {
         HealthComponent enemyHealth = healthMapper.get(enemy);
         int entityStrength = statsMapper.get(entity).getStrength();
         enemyHealth.setHealth(enemyHealth.getHealth() - entityStrength);
+
         if (spriterMapper.has(entity)) {
             spriterMapper.get(entity).getSpriterAnimationController().attackEvent();
         }
+
         if (spriterMapper.has(enemy)) {
-            SpriterModelComponent.SpriterAnimationController animationController = spriterMapper.get(enemy).getSpriterAnimationController();
+            SpriterAnimationController animationController = spriterMapper.get(enemy).getSpriterAnimationController();
+            Animation animation = animationController.getPlayer().getAnimation();
             if (enemyHealth.getHealth() <= 0) {
                 animationController.dieEvent();
-            } else {
+            } else if (animation.id == AnimationID.DAMAGED.getId()){
                 animationController.damagedEvent();
             }
         }
+
         entity.remove(AttackComponent.class);
     }
 }
