+++ desktop/src/com/karkoon/dungeoncrawler/desktop/DesktopLauncher.java
@@ -7,9 +7,9 @@ import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 public class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		System.setProperty("org.lwjgl.opengl.Window.undecorated", "true");
+/*		System.setProperty("org.lwjgl.opengl.Window.undecorated", "true");
         config.fullscreen = false;
-		config.resizable = false;
+		config.resizable = false;*/
         config.width = LwjglApplicationConfiguration.getDesktopDisplayMode().width;
         config.height = LwjglApplicationConfiguration.getDesktopDisplayMode().height;
 		new LwjglApplication(new DungeonCrawler(), config);
